Splitting up conversion and checking in to their own separate modules
